requirements.txt
Django==1.11.29
django-crispy-forms==1.9.0
mysqlclient==1.4.6
pytz==2020.5
